Datepicker became datetimepicker.
The datepicker has now buttons for selecting common times
- midnight, 6am, noon as well as a close button.
The time the button selects is determined by the class
"js-timepick js-time-HHMM".
Clicking outside the datetimepicker closes it now.
Picking date for an empty input defaults the time to 00:02.
Picking time for an empty input defaults the date to the next 24 hours.