JavaScript redirects seem to work as desired. Fixes #2923
Callbacks to load_content are now called after the request
is processed, even if it redirects.
Redirecting requests now chain correctly in request queue.

load_content now accepts boolean option is_priority_request.
When set to true, it prepends the request to the queue instead
of appending it. Should only be used internally for redirects.

New function alloc_loadbuf(is_priority?) returns index to
LOAD_BUF for a new request. If cannot make a priority request,
returns an index as for a normal one. A priority request can
only be made when the queue is empty (in which case priority
doesn't matter) or when a request has already been processed
and the queue thus has free slots on the start.

Error callback is now called when the request succeeds but target
element is not found.

Callbacks to load_content now get no parameters but get the object
passed to load_content as 'this'.