Photo mass upload now redirects correctly after save; fixes #26
In magnifier-glass add-object overlays,
a button that pushes to address stack (NewmanLib.ADR_STACK) can now have the
'js-custom-adrstack-callbacks'. This will prevent the default callbacks to be
attached.
(those save the ID of the added object and insert it to the original form)

The button that opens mass-upload now uses this special-casing mechanism.

The mass-upload onAllComplete event now calls newman's standard success-handler
instead of just redirecting to /photos/photo/