Newman's address stack now pushes correctly; fixes #66
The handler for .js-hashadr now saves the original address in the click event
and the handler for .js-adrstack-push now checks this before using the current
address for its `from' field.

This addresses the case when the document is already in the new address after
clicking a link and the address stack has no way of figuring out where we
came from.