Mass uploader can now default-populate unique photo fields; fixes #82
New fields for title and description of *all* mass-uploaded photos are now
shown. They are copied over to all photos without the corresponding fields
filled with a number appended.
The number is increased for every photo where at least one empty unique field
occurs. The number is constant for all fields of a given photo.